+++ src/3000.h
@@ -12,7 +12,7 @@ extern int ULTIMA_35ec_SelectDirection(void);
 extern int ULTIMA_368e_FindNpcTileAtPos(int x, int y, int z);
 extern int ULTIMA_3702(int param_1, int param_2, int param_3);
 extern int ULTIMA_38e4(void);
-extern void ULTIMA_39cc_SetTile(int new_tile_id, byte x, byte y);
+extern void ULTIMA_39cc_SetTile(byte new_tile_id, byte x, byte y);
 extern int ULTIMA_39fc_GetFirstActivePartyMember(void);
 extern void ULTIMA_3a74(byte a, byte b, byte c, byte d, byte e, byte f, int g);
 extern int ULTIMA_3aae(int a);
